requirements.txt
streamlit==1.29.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
Pillow==10.1.0
